src: move all icons into icons folder
