+++ src/components/AboutSection.js
@@ -6,12 +6,12 @@ import List from "@material-ui/core/List"
 import ListItem from "@material-ui/core/ListItem"
 import ListItemIcon from "@material-ui/core/ListItemIcon"
 import ListItemText from "@material-ui/core/ListItemText"
-import GithubIcon from "./GithubIcon"
-import TwitterIcon from "./TwitterIcon"
-import LinkedinIcon from "./LinkedinIcon"
-import FacebookIcon from "./FacebookIcon"
-import PhoneIcon from "./PhoneIcon"
-import MailIcon from "./MailIcon"
+import GithubIcon from "../icons/GithubIcon"
+import TwitterIcon from "../icons/TwitterIcon"
+import LinkedinIcon from "../icons/LinkedinIcon"
+import FacebookIcon from "../icons/FacebookIcon"
+import PhoneIcon from "../icons/PhoneIcon"
+import MailIcon from "../icons/MailIcon"
 import Section from "./Section"
 
 const styles = theme => ({
