+++ src/icons/LinkedinIcon.js
